review_paper_writing_crew_new/requirements.txt
crewai>=0.28.0
pinecone>=6.0.0
openai>=1.3.0
python-dotenv>=1.0.0
langchain>=0.0.335
tiktoken>=0.5.1 